+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- page navigations on the favorites view
+
 ### Changed
-* deb config
+- deb config
 
 ## [0.1.0] (2025-10-22)
 ### Added